There's no def-main in Python. That's C++.
FYI: This is a much prettier way to execute something that we want
to be executed on running the file. Defining a main() function is
what you would do in C++ (maybe C?) because it's defined to run on
execution on the program. So it's pretty useless in Python.

This checks if __name__ equals __main__ which would mean that we're
specifically running this file (as in `python tictactoe3d.py`).